# Map: Optimize away `isinst` check