Following a discussion here https://github.com/dotnet/fsharp/pull/10768

Found a simple perf gain at the cost of 2 bytes per a Map item. A significant gain in `getItem/contains/add` due to replacement of `isinst` by `int32` equality check. Benchmarked against current main.

The change is to store the height in leaves and share the field with nodes. Use it as an implicit tag (a leaf when `height = 1`) instead of the type check. Compared to the old discussion, when Left/Right were proposed to be stored in a universal node, this adds 4 bytes to leaves or 2 bytes per item on average (vs 16/8).

The tradeoff is basically the only thing to consider. Code changes are trivial.

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
Intel Core i7-8700 CPU 3.20GHz (Coffee Lake), 1 CPU, 12 logical and 6 physical cores
.NET Core SDK=5.0.200-preview.20601.7
  [Host] : .NET Core 5.0.1 (CoreCLR 5.0.120.57516, CoreFX 5.0.120.57516), X64 RyuJIT DEBUG
  After  : .NET Core 5.0.1 (CoreCLR 5.0.120.57516, CoreFX 5.0.120.57516), X64 RyuJIT
  Before : .NET Core 5.0.1 (CoreCLR 5.0.120.57516, CoreFX 5.0.120.57516), X64 RyuJIT

MaxRelativeError=0.01  Arguments=/p:Optimize=true  IterationCount=20  
IterationTime=250.0000 ms  WarmupCount=1  

```
|      Method |    Job | BuildConfiguration |  Size |          Mean |        Error |       StdDev | Rank |    Gen 0 |  Gen 1 | Gen 2 | Allocated | Code Size |
|------------ |------- |------------------- |------ |--------------:|-------------:|-------------:|-----:|---------:|-------:|------:|----------:|----------:|
|     **getItem** |  **After** |         **LocalBuild** |   **100** |      **28.34 ns** |     **0.291 ns** |     **0.324 ns** |    **1** |        **-** |      **-** |     **-** |         **-** |      **93 B** |
|     getItem | Before |            Default |   100 |      36.48 ns |     0.206 ns |     0.212 ns |    2 |        - |      - |     - |         - |     119 B |
|     **getItem** |  **After** |         **LocalBuild** | **10000** |      **60.40 ns** |     **0.365 ns** |     **0.421 ns** |    **3** |        **-** |      **-** |     **-** |         **-** |      **93 B** |
|     getItem | Before |            Default | 10000 |      82.42 ns |     0.446 ns |     0.513 ns |    4 |        - |      - |     - |         - |     119 B |
|             |        |                    |       |               |              |              |      |          |        |       |           |           |
| **containsKey** |  **After** |         **LocalBuild** |   **100** |      **24.95 ns** |     **0.231 ns** |     **0.266 ns** |    **1** |        **-** |      **-** |     **-** |         **-** |     **155 B** |
| containsKey | Before |            Default |   100 |      29.23 ns |     0.385 ns |     0.412 ns |    2 |        - |      - |     - |         - |     175 B |
| **containsKey** |  **After** |         **LocalBuild** | **10000** |      **46.19 ns** |     **0.292 ns** |     **0.337 ns** |    **3** |        **-** |      **-** |     **-** |         **-** |     **155 B** |
| containsKey | Before |            Default | 10000 |      53.00 ns |     0.177 ns |     0.196 ns |    4 |        - |      - |     - |         - |     175 B |
|             |        |                    |       |               |              |              |      |          |        |       |           |           |
|   **itemCount** |  **After** |         **LocalBuild** |   **100** |     **179.55 ns** |     **1.426 ns** |     **1.643 ns** |    **1** |        **-** |      **-** |     **-** |         **-** |      **81 B** |
|   itemCount | Before |            Default |   100 |     217.68 ns |     1.816 ns |     1.865 ns |    2 |        - |      - |     - |         - |      96 B |
|   **itemCount** |  **After** |         **LocalBuild** | **10000** |  **33,454.41 ns** | **1,561.752 ns** | **1,798.516 ns** |    **3** |        **-** |      **-** |     **-** |         **-** |      **81 B** |
|   itemCount | Before |            Default | 10000 |  34,357.14 ns |   374.999 ns |   431.850 ns |    3 |        - |      - |     - |         - |      96 B |
|             |        |                    |       |               |              |              |      |          |        |       |           |           |
| **iterForeach** |  **After** |         **LocalBuild** |   **100** |   **3,223.95 ns** |    **23.071 ns** |    **23.692 ns** |    **2** |   **1.0311** |      **-** |     **-** |    **6520 B** |     **283 B** |
| iterForeach | Before |            Default |   100 |   3,082.44 ns |    28.212 ns |    30.186 ns |    1 |   0.9704 |      - |     - |    6120 B |     283 B |
| **iterForeach** |  **After** |         **LocalBuild** | **10000** | **336,049.08 ns** | **1,375.830 ns** | **1,472.123 ns** |    **4** | **101.0638** |      **-** |     **-** |  **640120 B** |     **283 B** |
| iterForeach | Before |            Default | 10000 | 317,291.41 ns |   511.972 ns |   525.758 ns |    3 |  95.6633 |      - |     - |  600120 B |     283 B |
|             |        |                    |       |               |              |              |      |          |        |       |           |           |
|     **addItem** |  **After** |         **LocalBuild** |   **100** |     **145.75 ns** |     **0.811 ns** |     **0.934 ns** |    **1** |   **0.0590** |      **-** |     **-** |     **374 B** |     **601 B** |
|     addItem | Before |            Default |   100 |     170.09 ns |     2.717 ns |     2.907 ns |    2 |   0.0584 |      - |     - |     369 B |     603 B |
|     **addItem** |  **After** |         **LocalBuild** | **10000** |  **36,414.43 ns** |   **415.002 ns** |   **477.917 ns** |    **3** |  **11.0000** | **3.6250** |     **-** |   **69724 B** |     **601 B** |
|     addItem | Before |            Default | 10000 |  37,677.58 ns |   103.857 ns |   111.126 ns |    4 |  11.0000 | 3.2500 |     - |   69324 B |     603 B |
|             |        |                    |       |               |              |              |      |          |        |       |           |           |
|  **removeItem** |  **After** |         **LocalBuild** |   **100** |      **12.33 ns** |     **0.114 ns** |     **0.112 ns** |    **1** |   **0.0064** |      **-** |     **-** |      **40 B** |     **425 B** |
|  removeItem | Before |            Default |   100 |      12.22 ns |     0.076 ns |     0.081 ns |    1 |   0.0064 |      - |     - |      40 B |     443 B |
|  **removeItem** |  **After** |         **LocalBuild** | **10000** |   **1,189.82 ns** |     **6.181 ns** |     **7.118 ns** |    **2** |   **0.6345** |      **-** |     **-** |    **4000 B** |     **425 B** |
|  removeItem | Before |            Default | 10000 |   1,188.84 ns |     7.846 ns |     8.720 ns |    2 |   0.6345 |      - |     - |    4000 B |     443 B |