@buybackoff 2021-01-07
Initially (in the benchmarks table) in the `asNode` function I used `(# "" value: MapTreeNode<'Key,'Value> #)` instead of `value :?> MapTreeNode<'Key,'Value>`. This is a replacement for `Unsafe.As`, discussed long time ago [here](https://github.com/dotnet/fsharp/pull/5360#issuecomment-496831965).

But that fails in a bad way on Windows/full framework. The updated benchmarks are below:

There is still an improvement, but smaller. Making this a draft.

Is there a way to suppress the error for `(# "" value: MapTreeNode<'Key,'Value> #)` or achieve the same thing without a cast and without `S.R.CS.Unsafe` dependency? 


``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
Intel Core i7-8700 CPU 3.20GHz (Coffee Lake), 1 CPU, 12 logical and 6 physical cores
.NET Core SDK=5.0.200-preview.20601.7
  [Host] : .NET Core 5.0.1 (CoreCLR 5.0.120.57516, CoreFX 5.0.120.57516), X64 RyuJIT DEBUG
  After  : .NET Core 5.0.1 (CoreCLR 5.0.120.57516, CoreFX 5.0.120.57516), X64 RyuJIT
  Before : .NET Core 5.0.1 (CoreCLR 5.0.120.57516, CoreFX 5.0.120.57516), X64 RyuJIT

MaxRelativeError=0.01  Arguments=/p:Optimize=true  IterationCount=20  
IterationTime=250.0000 ms  WarmupCount=1  

```
|      Method |    Job | BuildConfiguration |  Size |          Mean |        Error |       StdDev | Rank |    Gen 0 |  Gen 1 | Gen 2 | Allocated | Code Size |
|------------ |------- |------------------- |------ |--------------:|-------------:|-------------:|-----:|---------:|-------:|------:|----------:|----------:|
|     **getItem** |  **After** |         **LocalBuild** |   **100** |      **34.16 ns** |     **0.258 ns** |     **0.297 ns** |    **1** |        **-** |      **-** |     **-** |         **-** |      **93 B** |
|     getItem | Before |            Default |   100 |      36.51 ns |     0.291 ns |     0.312 ns |    2 |        - |      - |     - |         - |     119 B |
|     **getItem** |  **After** |         **LocalBuild** | **10000** |      **75.90 ns** |     **0.839 ns** |     **0.933 ns** |    **3** |        **-** |      **-** |     **-** |         **-** |      **93 B** |
|     getItem | Before |            Default | 10000 |      82.41 ns |     0.306 ns |     0.352 ns |    4 |        - |      - |     - |         - |     119 B |
|             |        |                    |       |               |              |              |      |          |        |       |           |           |
| **containsKey** |  **After** |         **LocalBuild** |   **100** |      **29.02 ns** |     **0.310 ns** |     **0.357 ns** |    **1** |        **-** |      **-** |     **-** |         **-** |     **198 B** |
| containsKey | Before |            Default |   100 |      28.70 ns |     0.308 ns |     0.355 ns |    1 |        - |      - |     - |         - |     175 B |
| **containsKey** |  **After** |         **LocalBuild** | **10000** |      **52.57 ns** |     **0.453 ns** |     **0.465 ns** |    **2** |        **-** |      **-** |     **-** |         **-** |     **198 B** |
| containsKey | Before |            Default | 10000 |      53.28 ns |     0.423 ns |     0.487 ns |    2 |        - |      - |     - |         - |     175 B |
|             |        |                    |       |               |              |              |      |          |        |       |           |           |
|   **itemCount** |  **After** |         **LocalBuild** |   **100** |     **205.14 ns** |     **1.642 ns** |     **1.891 ns** |    **1** |        **-** |      **-** |     **-** |         **-** |     **110 B** |
|   itemCount | Before |            Default |   100 |     216.20 ns |     0.965 ns |     1.072 ns |    2 |        - |      - |     - |         - |      96 B |
|   **itemCount** |  **After** |         **LocalBuild** | **10000** |  **35,507.48 ns** |   **934.860 ns** | **1,076.586 ns** |    **4** |        **-** |      **-** |     **-** |         **-** |     **110 B** |
|   itemCount | Before |            Default | 10000 |  34,428.44 ns |   534.973 ns |   616.076 ns |    3 |        - |      - |     - |         - |      96 B |
|             |        |                    |       |               |              |              |      |          |        |       |           |           |
| **iterForeach** |  **After** |         **LocalBuild** |   **100** |   **3,139.77 ns** |    **15.790 ns** |    **18.184 ns** |    **2** |   **1.0354** |      **-** |     **-** |    **6520 B** |     **283 B** |
| iterForeach | Before |            Default |   100 |   3,065.72 ns |    19.452 ns |    22.401 ns |    1 |   0.9659 |      - |     - |    6120 B |     283 B |
| **iterForeach** |  **After** |         **LocalBuild** | **10000** | **324,242.95 ns** | **1,222.667 ns** | **1,408.025 ns** |    **3** | **101.5625** |      **-** |     **-** |  **640120 B** |     **283 B** |
| iterForeach | Before |            Default | 10000 | 320,282.94 ns | 1,422.972 ns | 1,638.697 ns |    3 |  95.6633 |      - |     - |  600120 B |     283 B |
|             |        |                    |       |               |              |              |      |          |        |       |           |           |
|     **addItem** |  **After** |         **LocalBuild** |   **100** |     **151.75 ns** |     **0.757 ns** |     **0.810 ns** |    **1** |   **0.0591** |      **-** |     **-** |     **374 B** |     **655 B** |
|     addItem | Before |            Default |   100 |     168.32 ns |     0.436 ns |     0.502 ns |    2 |   0.0583 |      - |     - |     369 B |     603 B |
|     **addItem** |  **After** |         **LocalBuild** | **10000** |  **37,558.66 ns** |   **509.422 ns** |   **566.221 ns** |    **3** |  **11.0000** | **3.6250** |     **-** |   **69724 B** |     **655 B** |
|     addItem | Before |            Default | 10000 |  39,931.69 ns |   349.741 ns |   388.736 ns |    4 |  10.9375 | 3.2813 |     - |   69324 B |     603 B |
|             |        |                    |       |               |              |              |      |          |        |       |           |           |
|  **removeItem** |  **After** |         **LocalBuild** |   **100** |      **12.65 ns** |     **0.100 ns** |     **0.115 ns** |    **2** |   **0.0064** |      **-** |     **-** |      **40 B** |     **454 B** |
|  removeItem | Before |            Default |   100 |      12.35 ns |     0.147 ns |     0.169 ns |    1 |   0.0064 |      - |     - |      40 B |     443 B |
|  **removeItem** |  **After** |         **LocalBuild** | **10000** |   **1,290.84 ns** |    **21.354 ns** |    **22.848 ns** |    **4** |   **0.6351** |      **-** |     **-** |    **4000 B** |     **454 B** |
|  removeItem | Before |            Default | 10000 |   1,195.52 ns |     6.200 ns |     6.634 ns |    3 |   0.6346 |      - |     - |    4000 B |     443 B |