@buybackoff 2021-01-07
Results for `count = 10`:

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
Intel Core i7-8700 CPU 3.20GHz (Coffee Lake), 1 CPU, 12 logical and 6 physical cores
.NET Core SDK=5.0.200-preview.20601.7
  [Host] : .NET Core 5.0.1 (CoreCLR 5.0.120.57516, CoreFX 5.0.120.57516), X64 RyuJIT DEBUG
  After  : .NET Core 5.0.1 (CoreCLR 5.0.120.57516, CoreFX 5.0.120.57516), X64 RyuJIT
  Before : .NET Core 5.0.1 (CoreCLR 5.0.120.57516, CoreFX 5.0.120.57516), X64 RyuJIT

MaxRelativeError=0.01  Arguments=/p:Optimize=true  IterationCount=20  
IterationTime=250.0000 ms  WarmupCount=1  

```
|      Method |    Job | BuildConfiguration | Size |        Mean |    Error |   StdDev | Rank |  Gen 0 | Gen 1 | Gen 2 | Allocated | Code Size |
|------------ |------- |------------------- |----- |------------:|---------:|---------:|-----:|-------:|------:|------:|----------:|----------:|
|     getItem |  After |         LocalBuild |   10 |    17.92 ns | 0.055 ns | 0.057 ns |    1 |      - |     - |     - |         - |      93 B |
|     getItem | Before |            Default |   10 |    36.07 ns | 0.329 ns | 0.379 ns |    2 |      - |     - |     - |         - |     119 B |
|             |        |                    |      |             |          |          |      |        |       |       |           |           |
| containsKey |  After |         LocalBuild |   10 |    16.03 ns | 0.117 ns | 0.134 ns |    1 |      - |     - |     - |         - |     198 B |
| containsKey | Before |            Default |   10 |    28.26 ns | 0.241 ns | 0.268 ns |    2 |      - |     - |     - |         - |     175 B |
|             |        |                    |      |             |          |          |      |        |       |       |           |           |
|   itemCount |  After |         LocalBuild |   10 |    19.27 ns | 0.170 ns | 0.189 ns |    1 |      - |     - |     - |         - |     110 B |
|   itemCount | Before |            Default |   10 |   217.36 ns | 1.386 ns | 1.540 ns |    2 |      - |     - |     - |         - |      96 B |
|             |        |                    |      |             |          |          |      |        |       |       |           |           |
| iterForeach |  After |         LocalBuild |   10 |   336.97 ns | 0.816 ns | 0.801 ns |    1 | 0.1201 |     - |     - |     760 B |     283 B |
| iterForeach | Before |            Default |   10 | 3,045.81 ns | 7.461 ns | 8.293 ns |    2 | 0.9686 |     - |     - |    6120 B |     283 B |
|             |        |                    |      |             |          |          |      |        |       |       |           |           |
|     addItem |  After |         LocalBuild |   10 |    69.05 ns | 0.414 ns | 0.477 ns |    1 | 0.0313 |     - |     - |     198 B |     655 B |
|     addItem | Before |            Default |   10 |   172.04 ns | 0.954 ns | 1.061 ns |    2 | 0.0587 |     - |     - |     369 B |     603 B |
|             |        |                    |      |             |          |          |      |        |       |       |           |           |
|  removeItem |  After |         LocalBuild |   10 |    14.13 ns | 0.060 ns | 0.066 ns |    2 | 0.0070 |     - |     - |      44 B |     454 B |
|  removeItem | Before |            Default |   10 |    12.13 ns | 0.043 ns | 0.046 ns |    1 | 0.0064 |     - |     - |      40 B |     443 B |