@buybackoff 2021-01-07
Updated benchmarks after changing `match` to `if`

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
Intel Core i7-8700 CPU 3.20GHz (Coffee Lake), 1 CPU, 12 logical and 6 physical cores
.NET Core SDK=5.0.200-preview.20601.7
  [Host] : .NET Core 5.0.1 (CoreCLR 5.0.120.57516, CoreFX 5.0.120.57516), X64 RyuJIT DEBUG
  After  : .NET Core 5.0.1 (CoreCLR 5.0.120.57516, CoreFX 5.0.120.57516), X64 RyuJIT
  Before : .NET Core 5.0.1 (CoreCLR 5.0.120.57516, CoreFX 5.0.120.57516), X64 RyuJIT

MaxRelativeError=0.01  Arguments=/p:Optimize=true  IterationCount=20  
IterationTime=250.0000 ms  WarmupCount=1  

```
|      Method |    Job | BuildConfiguration |  Size |          Mean |        Error |       StdDev | Rank |    Gen 0 |  Gen 1 | Gen 2 | Allocated | Code Size |
|------------ |------- |------------------- |------ |--------------:|-------------:|-------------:|-----:|---------:|-------:|------:|----------:|----------:|
|     **getItem** |  **After** |         **LocalBuild** |   **100** |      **28.65 ns** |     **0.381 ns** |     **0.424 ns** |    **1** |        **-** |      **-** |     **-** |         **-** |      **93 B** |
|     getItem | Before |            Default |   100 |      36.39 ns |     0.226 ns |     0.251 ns |    2 |        - |      - |     - |         - |     119 B |
|     **getItem** |  **After** |         **LocalBuild** | **10000** |      **61.33 ns** |     **0.263 ns** |     **0.293 ns** |    **3** |        **-** |      **-** |     **-** |         **-** |      **93 B** |
|     getItem | Before |            Default | 10000 |      82.86 ns |     0.129 ns |     0.143 ns |    4 |        - |      - |     - |         - |     119 B |
|             |        |                    |       |               |              |              |      |          |        |       |           |           |
| **containsKey** |  **After** |         **LocalBuild** |   **100** |      **27.50 ns** |     **0.256 ns** |     **0.285 ns** |    **1** |        **-** |      **-** |     **-** |         **-** |     **200 B** |
| containsKey | Before |            Default |   100 |      29.29 ns |     0.342 ns |     0.380 ns |    2 |        - |      - |     - |         - |     175 B |
| **containsKey** |  **After** |         **LocalBuild** | **10000** |      **53.51 ns** |     **0.289 ns** |     **0.309 ns** |    **3** |        **-** |      **-** |     **-** |         **-** |     **200 B** |
| containsKey | Before |            Default | 10000 |      54.77 ns |     0.341 ns |     0.393 ns |    4 |        - |      - |     - |         - |     175 B |
|             |        |                    |       |               |              |              |      |          |        |       |           |           |
|   **itemCount** |  **After** |         **LocalBuild** |   **100** |     **197.96 ns** |     **1.155 ns** |     **1.284 ns** |    **1** |        **-** |      **-** |     **-** |         **-** |     **112 B** |
|   itemCount | Before |            Default |   100 |     217.04 ns |     1.008 ns |     1.079 ns |    2 |        - |      - |     - |         - |      96 B |
|   **itemCount** |  **After** |         **LocalBuild** | **10000** |  **34,628.61 ns** |   **651.610 ns** |   **750.396 ns** |    **3** |        **-** |      **-** |     **-** |         **-** |     **112 B** |
|   itemCount | Before |            Default | 10000 |  34,815.34 ns | 1,108.432 ns | 1,186.009 ns |    3 |        - |      - |     - |         - |      96 B |
|             |        |                    |       |               |              |              |      |          |        |       |           |           |
| **iterForeach** |  **After** |         **LocalBuild** |   **100** |   **3,090.08 ns** |    **10.900 ns** |    **11.663 ns** |    **1** |   **1.0373** |      **-** |     **-** |    **6520 B** |     **283 B** |
| iterForeach | Before |            Default |   100 |   3,057.11 ns |    10.379 ns |    11.537 ns |    1 |   0.9728 |      - |     - |    6120 B |     283 B |
| **iterForeach** |  **After** |         **LocalBuild** | **10000** | **324,417.64 ns** | **1,901.370 ns** | **2,189.621 ns** |    **2** | **102.0408** |      **-** |     **-** |  **640120 B** |     **283 B** |
| iterForeach | Before |            Default | 10000 | 322,203.22 ns | 1,951.415 ns | 2,247.253 ns |    2 |  95.6633 |      - |     - |  600120 B |     283 B |
|             |        |                    |       |               |              |              |      |          |        |       |           |           |
|     **addItem** |  **After** |         **LocalBuild** |   **100** |     **153.87 ns** |     **1.593 ns** |     **1.834 ns** |    **1** |   **0.0593** |      **-** |     **-** |     **374 B** |     **668 B** |
|     addItem | Before |            Default |   100 |     168.03 ns |     0.663 ns |     0.763 ns |    2 |   0.0583 |      - |     - |     369 B |     603 B |
|     **addItem** |  **After** |         **LocalBuild** | **10000** |  **39,308.48 ns** | **1,460.013 ns** | **1,562.198 ns** |    **3** |  **11.0938** | **3.7500** |     **-** |   **69724 B** |     **668 B** |
|     addItem | Before |            Default | 10000 |  40,129.25 ns |   179.695 ns |   192.272 ns |    3 |  10.9375 | 3.2813 |     - |   69324 B |     603 B |
|             |        |                    |       |               |              |              |      |          |        |       |           |           |
|  **removeItem** |  **After** |         **LocalBuild** |   **100** |      **12.99 ns** |     **0.067 ns** |     **0.077 ns** |    **2** |   **0.0064** |      **-** |     **-** |      **40 B** |     **461 B** |
|  removeItem | Before |            Default |   100 |      12.20 ns |     0.074 ns |     0.085 ns |    1 |   0.0064 |      - |     - |      40 B |     443 B |
|  **removeItem** |  **After** |         **LocalBuild** | **10000** |   **1,292.43 ns** |    **16.367 ns** |    **17.512 ns** |    **4** |   **0.6358** |      **-** |     **-** |    **4000 B** |     **461 B** |
|  removeItem | Before |            Default | 10000 |   1,208.58 ns |     6.424 ns |     6.597 ns |    3 |   0.6347 |      - |     - |    4000 B |     443 B |